clippy happy

+++ crates/shadowsocks/src/wg/device/kqueue.rs
@@ -21,10 +21,10 @@ pub enum WaitResult<'a, H> {
 
 /// Implements a registry of pollable events
 pub struct EventPoll<H: Sized> {
-    events: Mutex<Vec<Option<Box<Event<H>>>>>, // Events with a file descriptor
-    custom: Mutex<Vec<Option<Box<Event<H>>>>>, // Other events (i.e. timers & notifiers)
+    events: Mutex<Vec<Option<Box<Event<H>>>>>,  // Events with a file descriptor
+    custom: Mutex<Vec<Option<Box<Event<H>>>>>,  // Other events (i.e. timers & notifiers)
     signals: Mutex<Vec<Option<Box<Event<H>>>>>, // Signal handlers
-    kqueue: RawFd,                             // The OS kqueue
+    kqueue: RawFd,                              // The OS kqueue
 }
 
 /// A type that hold a reference to a triggered Event
@@ -214,7 +214,7 @@ impl<H: Send + Sync> EventPoll<H> {
         };
 
         let (trigger, index) = match ev.kind {
-            EventKind::FD | EventKind::Signal => (ev.event.ident as RawFd, ev.event.ident as usize),
+            EventKind::FD | EventKind::Signal => (ev.event.ident as RawFd, ev.event.ident),
             EventKind::Timer | EventKind::Notifier => (-(events.len() as RawFd) - 1, events.len()), // Custom events get negative identifiers, hopefully we will never have more than 2^31 events of each type
         };
 
@@ -290,7 +290,9 @@ impl<H: Send + Sync> EventPoll<H> {
 }
 
 impl<H> EventPoll<H> {
-    // This function is only safe to call when the event loop is not running
+    /// # Safety
+    ///
+    /// This function is only safe to call when the event loop is not running
     pub unsafe fn clear_event_by_fd(&self, index: RawFd) {
         let (mut events, index) = if index >= 0 {
             (self.events.lock(), index as usize)
